+++ src/Controller/TasksController.php
@@ -23,8 +23,7 @@ class TasksController extends AbstractController
         private ProjectRepository $projectRepository,
         private EntityManagerService $entityManagerService,
         private AvatarService $avatarService
-    ) {
-    }
+    ) {}
 
     /**
      * Affichage de toutes les tâches d'un projet.
@@ -34,6 +33,9 @@ class TasksController extends AbstractController
     {
         // On récupère le projet et les employés associés.
         $project = $this->entityManagerService->getEntity($this->projectRepository, $projectId);
+        if (!$project) {
+            throw $this->createNotFoundException('Le projet n\'existe pas.');
+        }
         $projectEmployees = $project->getEmployees()->toArray();
 
         // On génère les avatars des employés associés au projet.
@@ -60,6 +62,9 @@ class TasksController extends AbstractController
     {
         // On récupère le projet auquel la tâche va être associée.
         $project = $this->entityManagerService->getEntity($this->projectRepository, $projectId);
+        if (!$project) {
+            throw $this->createNotFoundException('Le projet n\'existe pas.');
+        }
 
         // On crée une nouvelle tâche et on renseigne le projet auquel elle est associée.
         $task = new Task();
@@ -93,9 +98,15 @@ class TasksController extends AbstractController
     {
         // On récupère le projet auquel la tâche est associée.
         $project = $this->entityManagerService->getEntity($this->projectRepository, $projectId);
+        if (!$project) {
+            throw $this->createNotFoundException('Le projet n\'existe pas.');
+        }
 
         // On récupère la tâche à mettre à jour.
         $task = $this->entityManagerService->getEntity($this->taskRepository, $taskId);
+        if (!$task) {
+            throw $this->createNotFoundException('La tâche n\'existe pas.');
+        }
 
         // On crée le formulaire et on ajoute les employés associés au projet en option du formulaire.
         $form = $this->createForm(TaskType::class, $task, [
@@ -124,6 +135,9 @@ class TasksController extends AbstractController
     public function deleteTask(int $projectId, int $taskId): Response
     {
         $task = $this->entityManagerService->getEntity($this->taskRepository, $taskId);
+        if (!$task) {
+            throw $this->createNotFoundException('La tâche n\'existe pas.');
+        }
 
         $this->entityManagerService->remove($task);
